@@ -16,15 +16,14 @@
 package net.objecthunter.exp4j;
 
 import net.objecthunter.exp4j.function.Function;
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
-import static org.junit.Assert.assertTrue;
 
-
-public class ExpressionValidateTest {
+class ExpressionValidateTest {
 
     /**
      * Dummy function with 2 arguments.
@@ -32,8 +31,8 @@ public class ExpressionValidateTest {
     private final Function beta = new Function("beta", 2) {
 
         @Override
-        public double apply(double... args) {
-            return args[1] - args[0];
+        public BigDecimal apply(final BigDecimal... args) {
+            return args[1].subtract(args[0]);
         }
     };
 
@@ -43,8 +42,8 @@ public class ExpressionValidateTest {
     private final Function gamma = new Function("gamma", 3) {
 
         @Override
-        public double apply(double... args) {
-            return args[0] * args[1] / args[2];
+        public BigDecimal apply(final BigDecimal... args) {
+            return args[0].multiply(args[1]).divide(args[2]);
         }
     };
 
@@ -54,10 +53,10 @@ public class ExpressionValidateTest {
     private final Function eta = new Function("eta", 7) {
 
         @Override
-        public double apply(double... args) {
-            double eta = 0;
-            for (double a : args) {
-                eta += a;
+        public BigDecimal apply(final BigDecimal... args) {
+            var eta = BigDecimal.ZERO;
+            for (final var a : args) {
+                eta = eta.add(a);
             }
             return eta;
         }
@@ -66,244 +65,244 @@ public class ExpressionValidateTest {
     // valid scenarios
 
     @Test
-    public void testValidateNumber() {
-        Expression exp = new ExpressionBuilder("1")
+    void testValidateNumber() {
+        final var exp = new ExpressionBuilder("1")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateNumberPositive() {
-        Expression exp = new ExpressionBuilder("+1")
+    void testValidateNumberPositive() {
+        final var exp = new ExpressionBuilder("+1")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateNumberNegative() {
-        Expression exp = new ExpressionBuilder("-1")
+    void testValidateNumberNegative() {
+        final var exp = new ExpressionBuilder("-1")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateOperator() {
-        Expression exp = new ExpressionBuilder("x + 1 + 2")
+    void testValidateOperator() {
+        final var exp = new ExpressionBuilder("x + 1 + 2")
                 .variable("x")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunction() {
-        Expression exp = new ExpressionBuilder("sin(x)")
+    void testValidateFunction() {
+        final var exp = new ExpressionBuilder("sin(x)")
                 .variable("x")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionPositive() {
-        Expression exp = new ExpressionBuilder("+sin(x)")
+    void testValidateFunctionPositive() {
+        final var exp = new ExpressionBuilder("+sin(x)")
                 .variable("x")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionNegative() {
-        Expression exp = new ExpressionBuilder("-sin(x)")
+    void testValidateFunctionNegative() {
+        final var exp = new ExpressionBuilder("-sin(x)")
                 .variable("x")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionAndOperator() {
-        Expression exp = new ExpressionBuilder("sin(x + 1 + 2)")
+    void testValidateFunctionAndOperator() {
+        final var exp = new ExpressionBuilder("sin(x + 1 + 2)")
                 .variable("x")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionWithTwoArguments() {
-        Expression exp = new ExpressionBuilder("beta(x, y)")
+    void testValidateFunctionWithTwoArguments() {
+        final var exp = new ExpressionBuilder("beta(x, y)")
                 .variables("x", "y")
                 .functions(beta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionWithTwoArgumentsAndOperator() {
-        Expression exp = new ExpressionBuilder("beta(x, y + 1)")
+    void testValidateFunctionWithTwoArgumentsAndOperator() {
+        final var exp = new ExpressionBuilder("beta(x, y + 1)")
                 .variables("x", "y")
                 .functions(beta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionWithThreeArguments() {
-        Expression exp = new ExpressionBuilder("gamma(x, y, z)")
+    void testValidateFunctionWithThreeArguments() {
+        final var exp = new ExpressionBuilder("gamma(x, y, z)")
                 .variables("x", "y", "z")
                 .functions(gamma)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionWithThreeArgumentsAndOperator() {
-        Expression exp = new ExpressionBuilder("gamma(x, y, z + 1)")
+    void testValidateFunctionWithThreeArgumentsAndOperator() {
+        final var exp = new ExpressionBuilder("gamma(x, y, z + 1)")
                 .variables("x", "y", "z")
                 .functions(gamma)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionWithTwoAndThreeArguments() {
-        Expression exp = new ExpressionBuilder("gamma(x, beta(y, h), z)")
+    void testValidateFunctionWithTwoAndThreeArguments() {
+        final var exp = new ExpressionBuilder("gamma(x, beta(y, h), z)")
                 .variables("x", "y", "z", "h")
                 .functions(gamma, beta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionWithTwoAndThreeArgumentsAndOperator() {
-        Expression exp = new ExpressionBuilder("gamma(x, beta(y, h), z + 1)")
+    void testValidateFunctionWithTwoAndThreeArgumentsAndOperator() {
+        final var exp = new ExpressionBuilder("gamma(x, beta(y, h), z + 1)")
                 .variables("x", "y", "z", "h")
                 .functions(gamma, beta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionWithTwoAndThreeArgumentsAndMultipleOperator() {
-        Expression exp = new ExpressionBuilder("gamma(x * 2 / 4, beta(y, h + 1 + 2), z + 1 + 2 + 3 + 4)")
+    void testValidateFunctionWithTwoAndThreeArgumentsAndMultipleOperator() {
+        final var exp = new ExpressionBuilder("gamma(x * 2 / 4, beta(y, h + 1 + 2), z + 1 + 2 + 3 + 4)")
                 .variables("x", "y", "z", "h")
                 .functions(gamma, beta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionWithSevenArguments() {
-        Expression exp = new ExpressionBuilder("eta(1, 2, 3, 4, 5, 6, 7)")
+    void testValidateFunctionWithSevenArguments() {
+        final var exp = new ExpressionBuilder("eta(1, 2, 3, 4, 5, 6, 7)")
                 .functions(eta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     @Test
-    public void testValidateFunctionWithSevenArgumentsAndOperator() {
-        Expression exp = new ExpressionBuilder("eta(1, 2, 3, 4, 5, 6, 7) * 2 * 3 * 4")
+    void testValidateFunctionWithSevenArgumentsAndOperator() {
+        final var exp = new ExpressionBuilder("eta(1, 2, 3, 4, 5, 6, 7) * 2 * 3 * 4")
                 .functions(eta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertTrue(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertTrue(result.isValid());
     }
 
     // invalid scenarios
 
     @Test
-    public void testValidateInvalidFunction() {
-        Expression exp = new ExpressionBuilder("sin()")
+    void testValidateInvalidFunction() {
+        final var exp = new ExpressionBuilder("sin()")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertFalse(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertFalse(result.isValid());
     }
 
     @Test
-    public void testValidateInvalidOperand() {
-        Expression exp = new ExpressionBuilder("1 + ")
+    void testValidateInvalidOperand() {
+        final var exp = new ExpressionBuilder("1 + ")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertFalse(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertFalse(result.isValid());
     }
 
     @Test
-    public void testValidateInvalidFunctionWithTooFewArguments() {
-        Expression exp = new ExpressionBuilder("beta(1)")
+    void testValidateInvalidFunctionWithTooFewArguments() {
+        final var exp = new ExpressionBuilder("beta(1)")
                 .functions(beta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertFalse(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertFalse(result.isValid());
     }
 
     @Test
-    public void testValidateInvalidFunctionWithTooFewArgumentsAndOperands() {
-        Expression exp = new ExpressionBuilder("beta(1 + )")
+    void testValidateInvalidFunctionWithTooFewArgumentsAndOperands() {
+        final var exp = new ExpressionBuilder("beta(1 + )")
                 .functions(beta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertFalse(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertFalse(result.isValid());
     }
 
     @Test
-    public void testValidateInvalidFunctionWithManyArguments() {
-        Expression exp = new ExpressionBuilder("beta(1, 2, 3)")
+    void testValidateInvalidFunctionWithManyArguments() {
+        final var exp = new ExpressionBuilder("beta(1, 2, 3)")
                 .functions(beta)
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertFalse(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertFalse(result.isValid());
     }
 
     @Test
-    public void testValidateInvalidOperator() {
-        Expression exp = new ExpressionBuilder("+")
+    void testValidateInvalidOperator() {
+        final var exp = new ExpressionBuilder("+")
                 .build();
-        ValidationResult result = exp.validate(false);
-        Assert.assertFalse(result.isValid());
+        final var result = exp.validate(false);
+        Assertions.assertFalse(result.isValid());
     }
 
     // Thanks go out to werwiesel for reporting the issue
     // https://github.com/fasseg/exp4j/issues/59
     @Test
-    public void testNoArgFunctionValidation() {
-        Function now = new Function("now", 0) {
+    void testNoArgFunctionValidation() {
+        final var now = new Function("now", 0) {
             @Override
-            public double apply(double... args) {
-                return (double) new Date().getTime();
+            public BigDecimal apply(final BigDecimal... args) {
+                return BigDecimal.valueOf(new Date().getTime());
             }
         };
-        Expression e = new ExpressionBuilder("14*now()")
+        var e = new ExpressionBuilder("14*now()")
                 .function(now)
                 .build();
-        assertTrue(e.validate().isValid());
+        Assertions.assertTrue(e.validate().isValid());
 
         e = new ExpressionBuilder("now()")
                 .function(now)
                 .build();
-        assertTrue(e.validate().isValid());
+        Assertions.assertTrue(e.validate().isValid());
 
         e = new ExpressionBuilder("sin(now())")
                 .function(now)
                 .build();
-        assertTrue(e.validate().isValid());
+        Assertions.assertTrue(e.validate().isValid());
 
         e = new ExpressionBuilder("sin(now()) % 14")
                 .function(now)
                 .build();
-        assertTrue(e.validate().isValid());
+        Assertions.assertTrue(e.validate().isValid());
     }
 
 }
