@@ -24,8 +24,8 @@ import net.objecthunter.exp4j.shuntingyard.ShuntingYard;
 import java.util.*;
 
 /**
- * Factory class for {@link Expression} instances. This class is the main API entrypoint. Users should create new
- * {@link Expression} instances using this factory class.
+ * Factory class for {@link Expression} instances. This class is the main API entrypoint. Users should create new {@link Expression}
+ * instances using this factory class.
  */
 public class ExpressionBuilder {
 
@@ -44,8 +44,8 @@ public class ExpressionBuilder {
      *
      * @param expression the expression to be parsed
      */
-    public ExpressionBuilder(String expression) {
-        if (expression == null || expression.trim().length() == 0) {
+    public ExpressionBuilder(final String expression) {
+        if (null == expression || expression.trim().isEmpty()) {
             throw new IllegalArgumentException("Expression can not be empty");
         }
         this.expression = expression;
@@ -57,10 +57,11 @@ public class ExpressionBuilder {
     /**
      * Add a {@link net.objecthunter.exp4j.function.Function} implementation available for use in the expression
      *
-     * @param function the custom {@link net.objecthunter.exp4j.function.Function} implementation that should be available for use in the expression.
+     * @param function the custom {@link net.objecthunter.exp4j.function.Function} implementation that should be available for use in the
+     *                 expression.
      * @return the ExpressionBuilder instance
      */
-    public ExpressionBuilder function(Function function) {
+    public ExpressionBuilder function(final Function function) {
         this.userFunctions.put(function.getName(), function);
         return this;
     }
@@ -71,8 +72,8 @@ public class ExpressionBuilder {
      * @param functions the custom {@link net.objecthunter.exp4j.function.Function} implementations
      * @return the ExpressionBuilder instance
      */
-    public ExpressionBuilder functions(Function... functions) {
-        for (Function f : functions) {
+    public ExpressionBuilder functions(final Function... functions) {
+        for (final var f : functions) {
             this.userFunctions.put(f.getName(), f);
         }
         return this;
@@ -84,8 +85,8 @@ public class ExpressionBuilder {
      * @param functions A {@link java.util.List} of custom {@link net.objecthunter.exp4j.function.Function} implementations
      * @return the ExpressionBuilder instance
      */
-    public ExpressionBuilder functions(List<Function> functions) {
-        for (Function f : functions) {
+    public ExpressionBuilder functions(final List<Function> functions) {
+        for (final var f : functions) {
             this.userFunctions.put(f.getName(), f);
         }
         return this;
@@ -97,7 +98,7 @@ public class ExpressionBuilder {
      * @param variableNames the variables used in the expression
      * @return the ExpressionBuilder instance
      */
-    public ExpressionBuilder variables(Set<String> variableNames) {
+    public ExpressionBuilder variables(final Set<String> variableNames) {
         this.variableNames.addAll(variableNames);
         return this;
     }
@@ -108,7 +109,7 @@ public class ExpressionBuilder {
      * @param variableNames the variables used in the expression
      * @return the ExpressionBuilder instance
      */
-    public ExpressionBuilder variables(String... variableNames) {
+    public ExpressionBuilder variables(final String... variableNames) {
         Collections.addAll(this.variableNames, variableNames);
         return this;
     }
@@ -119,12 +120,12 @@ public class ExpressionBuilder {
      * @param variableName the variable used in the expression
      * @return the ExpressionBuilder instance
      */
-    public ExpressionBuilder variable(String variableName) {
+    public ExpressionBuilder variable(final String variableName) {
         this.variableNames.add(variableName);
         return this;
     }
 
-    public ExpressionBuilder implicitMultiplication(boolean enabled) {
+    public ExpressionBuilder implicitMultiplication(final boolean enabled) {
         this.implicitMultiplication = enabled;
         return this;
     }
@@ -135,15 +136,15 @@ public class ExpressionBuilder {
      * @param operator the custom {@link net.objecthunter.exp4j.operator.Operator} to add
      * @return the ExpressionBuilder instance
      */
-    public ExpressionBuilder operator(Operator operator) {
+    public ExpressionBuilder operator(final Operator operator) {
         this.checkOperatorSymbol(operator);
         this.userOperators.put(operator.getSymbol(), operator);
         return this;
     }
 
-    private void checkOperatorSymbol(Operator op) {
-        String name = op.getSymbol();
-        for (char ch : name.toCharArray()) {
+    private void checkOperatorSymbol(final Operator op) {
+        final var name = op.getSymbol();
+        for (final var ch : name.toCharArray()) {
             if (!Operator.isAllowedOperatorChar(ch)) {
                 throw new IllegalArgumentException("The operator symbol '" + name + "' is invalid");
             }
@@ -156,8 +157,8 @@ public class ExpressionBuilder {
      * @param operators the set of custom {@link net.objecthunter.exp4j.operator.Operator} implementations to add
      * @return the ExpressionBuilder instance
      */
-    public ExpressionBuilder operator(Operator... operators) {
-        for (Operator o : operators) {
+    public ExpressionBuilder operator(final Operator... operators) {
+        for (final var o : operators) {
             this.operator(o);
         }
         return this;
@@ -169,8 +170,8 @@ public class ExpressionBuilder {
      * @param operators the {@link java.util.List} of custom {@link net.objecthunter.exp4j.operator.Operator} implementations to add
      * @return the ExpressionBuilder instance
      */
-    public ExpressionBuilder operator(List<Operator> operators) {
-        for (Operator o : operators) {
+    public ExpressionBuilder operator(final List<Operator> operators) {
+        for (final var o : operators) {
             this.operator(o);
         }
         return this;
@@ -182,7 +183,7 @@ public class ExpressionBuilder {
      * @return an {@link Expression} instance which can be used to evaluate the result of the expression
      */
     public Expression build() {
-        if (expression.length() == 0) {
+        if (expression.isEmpty()) {
             throw new IllegalArgumentException("The expression can not be empty");
         }
 
@@ -193,9 +194,9 @@ public class ExpressionBuilder {
         variableNames.add("φ");
 
         /* Check if there are duplicate vars/functions */
-        for (String var : variableNames) {
-            if (Functions.getBuiltinFunction(var) != null || userFunctions.containsKey(var)) {
-                throw new IllegalArgumentException("A variable can not have the same name as a function [" + var + "]");
+        for (final var variableName : variableNames) {
+            if (null != Functions.getBuiltinFunction(variableName) || userFunctions.containsKey(variableName)) {
+                throw new IllegalArgumentException("A variable can not have the same name as a function [" + variableName + "]");
             }
         }
 
