@@ -15,7 +15,14 @@
  */
 package net.objecthunter.exp4j.operator;
 
+import java.math.BigDecimal;
+import java.math.MathContext;
+import java.math.RoundingMode;
+
 public abstract class Operators {
+
+    public static final MathContext MC = new MathContext(34, RoundingMode.HALF_UP);
+
     private static final int INDEX_ADDITION = 0;
     private static final int INDEX_SUBTRACTION = 1;
     private static final int INDEX_MULTIPLICATION = 2;
@@ -30,56 +37,56 @@ public abstract class Operators {
     static {
         BUILT_IN_OPERATORS[INDEX_ADDITION] = new Operator("+", 2, true, Operator.PRECEDENCE_ADDITION) {
             @Override
-            public double apply(final double... args) {
-                return args[0] + args[1];
+            public BigDecimal apply(final BigDecimal... args) {
+                return args[0].add(args[1]);
             }
         };
         BUILT_IN_OPERATORS[INDEX_SUBTRACTION] = new Operator("-", 2, true, Operator.PRECEDENCE_ADDITION) {
             @Override
-            public double apply(final double... args) {
-                return args[0] - args[1];
+            public BigDecimal apply(final BigDecimal... args) {
+                return args[0].subtract(args[1]);
             }
         };
         BUILT_IN_OPERATORS[INDEX_UNARY_MINUS] = new Operator("-", 1, false, Operator.PRECEDENCE_UNARY_MINUS) {
             @Override
-            public double apply(final double... args) {
-                return -args[0];
+            public BigDecimal apply(final BigDecimal... args) {
+                return args[0].multiply(BigDecimal.valueOf(-1L));
             }
         };
         BUILT_IN_OPERATORS[INDEX_UNARY_PLUS] = new Operator("+", 1, false, Operator.PRECEDENCE_UNARY_PLUS) {
             @Override
-            public double apply(final double... args) {
+            public BigDecimal apply(final BigDecimal... args) {
                 return args[0];
             }
         };
         BUILT_IN_OPERATORS[INDEX_MULTIPLICATION] = new Operator("*", 2, true, Operator.PRECEDENCE_MULTIPLICATION) {
             @Override
-            public double apply(final double... args) {
-                return args[0] * args[1];
+            public BigDecimal apply(final BigDecimal... args) {
+                return args[0].multiply(args[1]);
             }
         };
         BUILT_IN_OPERATORS[INDEX_DIVISION] = new Operator("/", 2, true, Operator.PRECEDENCE_DIVISION) {
             @Override
-            public double apply(final double... args) {
-                if (args[1] == 0d) {
+            public BigDecimal apply(final BigDecimal... args) {
+                if (0 == args[1].signum()) {
                     throw new ArithmeticException("Division by zero!");
                 }
-                return args[0] / args[1];
+                return args[0].divide(args[1], MC);
             }
         };
         BUILT_IN_OPERATORS[INDEX_POWER] = new Operator("^", 2, false, Operator.PRECEDENCE_POWER) {
             @Override
-            public double apply(final double... args) {
-                return Math.pow(args[0], args[1]);
+            public BigDecimal apply(final BigDecimal... args) {
+                return BigDecimal.valueOf(Math.pow(args[0].doubleValue(), args[1].doubleValue()));
             }
         };
         BUILT_IN_OPERATORS[INDEX_MODULO] = new Operator("%", 2, true, Operator.PRECEDENCE_MODULO) {
             @Override
-            public double apply(final double... args) {
-                if (args[1] == 0d) {
+            public BigDecimal apply(final BigDecimal... args) {
+                if (0 == args[1].signum()) {
                     throw new ArithmeticException("Division by zero!");
                 }
-                return args[0] % args[1];
+                return BigDecimal.valueOf(args[0].doubleValue() % args[1].doubleValue());
             }
         };
     }
@@ -87,13 +94,13 @@ public abstract class Operators {
     public static Operator getBuiltinOperator(final char symbol, final int numArguments) {
         switch (symbol) {
             case '+':
-                if (numArguments != 1) {
+                if (1 != numArguments) {
                     return BUILT_IN_OPERATORS[INDEX_ADDITION];
                 }
 
                 return BUILT_IN_OPERATORS[INDEX_UNARY_PLUS];
             case '-':
-                if (numArguments != 1) {
+                if (1 != numArguments) {
                     return BUILT_IN_OPERATORS[INDEX_SUBTRACTION];
                 }
 
