@@ -21,6 +21,7 @@ import net.objecthunter.exp4j.operator.Operator;
  * Represents an operator used in expressions
  */
 public class OperatorToken extends Token {
+
     private final Operator operator;
 
     /**
@@ -28,9 +29,9 @@ public class OperatorToken extends Token {
      *
      * @param op the operator
      */
-    public OperatorToken(Operator op) {
+    public OperatorToken(final Operator op) {
         super(Token.TOKEN_OPERATOR);
-        if (op == null) {
+        if (null == op) {
             throw new IllegalArgumentException("Operator is unknown for token.");
         }
         this.operator = op;
